fixed some bugs

+++ src/main/java/com/onyshkiv/command/impl/admin/DeleteUserCommand.java
@@ -10,23 +10,30 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.Optional;
+
 public class DeleteUserCommand implements Command {
     private static final Logger logger = LogManager.getLogger(DeleteUserCommand.class);
     UserService userService = UserService.getInstance();
+
     @Override
     public CommandResult execute(HttpServletRequest req, HttpServletResponse resp) {
         String login = req.getParameter("login");
 
-        try{
-            User userToDeleting = userService.findUserByLogin(login).get();
+        try {
+            Optional<User> optionalUser = userService.findUserByLogin(login);
+            if (optionalUser.isEmpty()) {
+                logger.info(String.format("There are not user(librarian) with login %s", login));
+                return new CommandResult("/controller?action=getLibrarians", true);
+            }
+            User userToDeleting = optionalUser.get();
             userService.deleteUser(userToDeleting);
-            logger.info(String.format("Librarian %s has been deleted.",login));
-        }catch (ServiceException e){
-            logger.error("Problem with user service occurred!(#DeleteUserCommand->findUserByLogin)", e);
-            return new CommandResult("/controller?action=getLibrarians", true); //todo another redirect
-        }
-        //todo maybe some info, that user has been deleted (like in books.jsp when book added)
 
-        return new CommandResult("/controller?action=getLibrarians",true);
+        } catch (ServiceException e) {
+            logger.error("Problem with user service occurred!", e);
+            return new CommandResult("/controller?action=getLibrarians", true);
+        }
+        logger.info(String.format("Librarian %s has been deleted.", login));
+        return new CommandResult("/controller?action=getLibrarians", true);
     }
 }
